Refactors user-validate as message-outbound-validate
