+++ config/lib/middleware/messages/signup/message-outbound.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  messageDirection: 'outbound-api-send',
+  shouldPostToPlatform: true,
+  shouldSendWhilePaused: false,
+};
