+++ config/lib/middleware/messages/support/message-outbound.js
@@ -3,4 +3,5 @@
 module.exports = {
   messageDirection: 'outbound-api-send',
   shouldPostToPlatform: true,
+  shouldSendWhilePaused: true,
 };
