+++ lib/middleware/messages/message-outbound-validate.js
@@ -3,22 +3,14 @@
 const helpers = require('../../helpers');
 const UnprocessibleEntityError = require('../../../app/exceptions/UnprocessibleEntityError');
 
-module.exports = function validateUser() {
+module.exports = function validateOutbound(config) {
   return (req, res, next) => {
-    // This check is here because Front messages get routed through here.
-    // TODO: Remove this check once Send Message is split into v2 Front + Campaign Signup messages.
-    // It's safe to leave for Broadcast messages, because we have already checked for northstarId
-    // in our middleware/messages/broadcast/params.
-    if (!req.user) {
-      return next();
-    }
-
     if (!helpers.user.isSubscriber(req.user)) {
       const error = new UnprocessibleEntityError('Northstar User is unsubscribed.');
       return helpers.sendErrorResponseWithSuppressHeaders(res, error);
     }
 
-    if (helpers.user.isPaused(req.user)) {
+    if (config.shouldSendWhenPaused === false && helpers.user.isPaused(req.user)) {
       const error = new UnprocessibleEntityError('Northstar User conversation is paused.');
       return helpers.sendErrorResponseWithSuppressHeaders(res, error);
     }
