+++ lib/middleware/messages/user-get.js
@@ -5,15 +5,14 @@ const NotFoundError = require('../../../app/exceptions/NotFoundError');
 
 module.exports = function getNorthstarUser() {
   return (req, res, next) => {
-    // This check is here because Front messages get routed through here.
-    // TODO: Remove this check once Send Message is split into v2 Front + Campaign Signup messages.
-    // It's safe to leave for Broadcast messages, because we have already checked for northstarId
-    // in our middleware/messages/broadcast/params.
-    if (!req.userId) {
-      return next();
+    let promise;
+    if (req.userId) {
+      promise = helpers.user.fetchById(req.userId);
+    } else {
+      promise = req.conversation.getNorthstarUser();
     }
 
-    return helpers.user.fetchById(req.userId)
+    return promise
       .then((user) => {
         req.user = user;
